pyproject.toml
[project]
authors = [{ name = "Axelle LNU", email = "aaposramennoodles@gmail.com" }]
dependencies = [
    "nextcord>=2.5.0",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.0.1",
]
description = "An attempt to turn Hoyoverse characters into playable TCG characters like Genius Invocation TCG from Genshin Impact."
license = { file = "LICENSE" }
name = "aqua-tcg"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[tool.pyright]
enableTypeIgnoreComments = false
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportUnnecessaryIsInstance = true
reportUnnecessaryTypeIgnoreComment = true
typeCheckingMode = "standard"

[dependency-groups]
lint = ["ruff>=0.12.3"]
type-check = ["pyright[nodejs]>=1.1.403"]
